Bookkeeping

+++ scripts/condor_submit_nano_connect_ntuples.py
@@ -109,13 +109,14 @@ def write_sh(srcfile,ifile,ofile,logfile,outfile,errfile,dataset,filetag,n):
 
 if __name__ == "__main__":
     if not len(sys.argv) > 1 or '-h' in sys.argv or '--help' in sys.argv:
-        print "Usage: %s [-q queue] [-tree treename] [-list listfile.list] [-split S] [--sms] [--data] [--sys] [--dry-run] [--verbose]" % sys.argv[0]
+        print "Usage: %s [-q queue] [-tree treename] [-list listfile.list] [-split S] [--sms] [--data] [--sys] [--dry-run] [--verbose] [--count]" % sys.argv[0]
         sys.exit(1)
 
     argv_pos    = 1
     DO_SMS      = 0
     DO_DATA     = 0
     DRY_RUN     = 0
+    COUNT       = 0
     VERBOSE     = 0
     SYS         = 0
   
@@ -144,6 +145,9 @@ if __name__ == "__main__":
     if '--dry-run' in sys.argv:
         DRY_RUN = 1
         argv_pos += 1
+    if '--count' in sys.argv:
+        COUNT = 1
+        argv_pos += 1
     if '--verbose' in sys.argv:
         VERBOSE = 1
         argv_pos += 1
@@ -164,6 +168,9 @@ if __name__ == "__main__":
     if SYS:
         print " --- Processing SYS"
 
+    if COUNT:
+        print " --- Only Counting (No Processing)"
+
     # input sample list
     listfile = LIST
     listname = listfile.split("/")
@@ -173,69 +180,74 @@ if __name__ == "__main__":
     
     # create and organize output folders
     TARGET  = RUN_DIR+"/"+NAME+"/"
-    os.system("rm -rf "+TARGET)
-    os.system("mkdir -p "+TARGET)
+    if not COUNT:
+        os.system("rm -rf "+TARGET)
+        os.system("mkdir -p "+TARGET)
     listdir = TARGET+"list/"
     srcdir  = TARGET+"src/"
     logdir  = TARGET+"log/"
     outdir  = TARGET+"out/"
     errdir  = TARGET+"err/"
-    os.system("mkdir -p "+listdir)
-    os.system("mkdir -p "+logdir)
-    os.system("mkdir -p "+outdir)
-    os.system("mkdir -p "+errdir)
+    if not COUNT:
+        os.system("mkdir -p "+listdir)
+        os.system("mkdir -p "+logdir)
+        os.system("mkdir -p "+outdir)
+        os.system("mkdir -p "+errdir)
     os.system("mkdir -p "+srcdir)
 
     # make config directory
     config = TARGET+"config/"
-    os.system("mkdir -p "+config)
+    if not COUNT:
+        os.system("mkdir -p "+config)
 
     # NOTE: there is a bug for setting the hadd verbosity level, "hadd -v 0".
     # The hadd verbosity option only works in ROOT 6.18/00 and later.
     # https://github.com/root-project/root/issues/11372
     # https://github.com/root-project/root/pull/3914
 
-    # make EventCount file
-    os.system("hadd "+config+"EventCount.root root/EventCount/*.root > /dev/null")
-    EVTCNT = "./config/EventCount.root"
+    if not COUNT:
 
-    # make FilterEff file 
-    os.system("hadd "+config+"FilterEff.root root/FilterEff/*.root > /dev/null")
-    FILTEREFF = "./config/FilterEff.root"
+        # make EventCount file
+        os.system("hadd "+config+"EventCount.root root/EventCount/*.root > /dev/null")
+        EVTCNT = "./config/EventCount.root"
 
-    # make json file
-    os.system("cat json/GoodRunList/*.txt > "+config+"GRL_JSON.txt")
-    os.system("echo -n $(tr -d '\n' < "+config+"GRL_JSON.txt) > "+config+"GRL_JSON.txt")
-    JSON = "./config/GRL_JSON.txt"
+        # make FilterEff file 
+        os.system("hadd "+config+"FilterEff.root root/FilterEff/*.root > /dev/null")
+        FILTEREFF = "./config/FilterEff.root"
 
-    # copy PU root files
-    os.system("cp -r root/PU "+config+".")
-    PUFOLD = "./config/PU/"
+        # make json file
+        os.system("cat json/GoodRunList/*.txt > "+config+"GRL_JSON.txt")
+        os.system("echo -n $(tr -d '\n' < "+config+"GRL_JSON.txt) > "+config+"GRL_JSON.txt")
+        JSON = "./config/GRL_JSON.txt"
 
-    # copy BTAG SF files
-    os.system("cp -r root/BtagSF "+config+".")
-    os.system("cp -r csv/BtagSF/* "+config+"BtagSF/.")
-    BTAGFOLD = "./config/BtagSF/"
+        # copy PU root files
+        os.system("cp -r root/PU "+config+".")
+        PUFOLD = "./config/PU/"
 
-    # copy JME files
-    os.system("cp -r data/JME "+config+".")
-    JMEFOLD = "./config/JME/"
+        # copy BTAG SF files
+        os.system("cp -r root/BtagSF "+config+".")
+        os.system("cp -r csv/BtagSF/* "+config+"BtagSF/.")
+        BTAGFOLD = "./config/BtagSF/"
 
-    # copy MET trigger files
-    os.system("cp -r csv/METTrigger "+config+".")
-    METFILE = "./config/METTrigger/Parameters.csv"
+        # copy JME files
+        os.system("cp -r data/JME "+config+".")
+        JMEFOLD = "./config/JME/"
 
-    # copy Prefire files
-    os.system("cp -r root/Prefire "+config+".")
-    PREFIREFILE = "./config/Prefire/Prefire.root"
+        # copy MET trigger files
+        os.system("cp -r csv/METTrigger "+config+".")
+        METFILE = "./config/METTrigger/Parameters.csv"
 
-    # copy SV NN model
-    os.system("cat json/lwtnn/nano_train_model.json > "+config+"NNmodel.json")
-    SVFILE = "./config/NNmodel.json"
-    
-    os.system("cp "+EXE+" "+config+".")
-    os.system("cp "+RESTFRAMES+" "+config+".")
-    os.system("cp "+CMSSW_SETUP+" "+config+".")
+        # copy Prefire files
+        os.system("cp -r root/Prefire "+config+".")
+        PREFIREFILE = "./config/Prefire/Prefire.root"
+
+        # copy SV NN model
+        os.system("cat json/lwtnn/nano_train_model.json > "+config+"NNmodel.json")
+        SVFILE = "./config/NNmodel.json"
+        
+        os.system("cp "+EXE+" "+config+".")
+        os.system("cp "+RESTFRAMES+" "+config+".")
+        os.system("cp "+CMSSW_SETUP+" "+config+".")
 
     # output root files
     OUT_DIR = OUT_BASE+"/"+NAME+"/"
@@ -250,6 +262,7 @@ if __name__ == "__main__":
 
     knowntags = ["Fall17_94X","Autumn18_102X","Summer16_94X","Fall17_102X","Summer16_102X","Summer20UL16_102X","Summer20UL16APV_102X","Summer20UL17_102X","Summer20UL18_102X","RunIISummer20UL17NanoAODv9"]
     
+    n_samples = 0
     with open(listfile,'r') as mylist:
         inputlist = mylist.readlines()
 
@@ -265,6 +278,8 @@ if __name__ == "__main__":
             listfile = LIST
             listname = listfile.split("/")
             listname = listname[-1]
+
+            n_samples = n_samples+1
 
             dataset = flist.split("/")
             dataset = dataset[-1]
@@ -293,55 +308,60 @@ if __name__ == "__main__":
 
             if len(datasetlist) == 0:
                 datasetlist.append((dataset,filetag,rootlist))
-                os.system("rm -rf "+OUT_DIR+dataset+"_"+filetag+"/")
-                os.system("mkdir -p "+OUT_DIR+dataset+"_"+filetag+"/")
+                if not COUNT:
+                    os.system("rm -rf "+OUT_DIR+dataset+"_"+filetag+"/")
+                    os.system("mkdir -p "+OUT_DIR+dataset+"_"+filetag+"/")
                 continue
             
             tagtuple = [item for item in datasetlist if item[0] == dataset]
             if len(tagtuple) == 0:
                 datasetlist.append((dataset,filetag,rootlist))
-                os.system("rm -rf "+OUT_DIR+dataset+"_"+filetag+"/")
-                os.system("mkdir -p "+OUT_DIR+dataset+"_"+filetag+"/")
+                if not COUNT:
+                    os.system("rm -rf "+OUT_DIR+dataset+"_"+filetag+"/")
+                    os.system("mkdir -p "+OUT_DIR+dataset+"_"+filetag+"/")
                 continue
 
             p = datasetlist.index(tagtuple[0])
             datasetlist[p][2].extend(rootlist)
 
     for (dataset,filetag,rootlist) in datasetlist:
-        os.system("mkdir -p "+os.path.join(listdir, dataset+'_'+filetag))
-        listdir_sam = os.path.join(listdir, dataset+'_'+filetag)
-        listlist = create_filelist(rootlist, dataset, filetag)
-        overlist_name = listdir_sam+'/'+dataset+'_'+filetag+'_list.list'
-        with open(overlist_name,'w') as overlist:
-            newlistlist = ['config/'+'/'.join(l.split('/')[-3:])+'\n' for l in listlist]
-            overlist.writelines(newlistlist)
-            overlist.close()
+        if not COUNT:
+            os.system("mkdir -p "+os.path.join(listdir, dataset+'_'+filetag))
+            listdir_sam = os.path.join(listdir, dataset+'_'+filetag)
+            listlist = create_filelist(rootlist, dataset, filetag)
+            overlist_name = listdir_sam+'/'+dataset+'_'+filetag+'_list.list'
+            with open(overlist_name,'w') as overlist:
+                newlistlist = ['config/'+'/'.join(l.split('/')[-3:])+'\n' for l in listlist]
+                overlist.writelines(newlistlist)
+                overlist.close()
 
-        os.system("mkdir -p "+os.path.join(logdir, dataset+'_'+filetag))
-        os.system("mkdir -p "+os.path.join(outdir, dataset+'_'+filetag))
-        os.system("mkdir -p "+os.path.join(errdir, dataset+'_'+filetag))
+        if not COUNT:
+            os.system("mkdir -p "+os.path.join(logdir, dataset+'_'+filetag))
+            os.system("mkdir -p "+os.path.join(outdir, dataset+'_'+filetag))
+            os.system("mkdir -p "+os.path.join(errdir, dataset+'_'+filetag))
 
-        file_name = os.path.join(OUT_DIR, dataset+'_'+filetag, overlist_name.split('/')[-1].replace('_list.list', '_$(ItemIndex)_$(Step)'))
+            file_name = os.path.join(OUT_DIR, dataset+'_'+filetag, overlist_name.split('/')[-1].replace('_list.list', '_$(ItemIndex)_$(Step)'))
 
-        logfile = os.path.join(logdir, dataset+'_'+filetag, file_name.split('/')[-1])
-        outfile = os.path.join(outdir, dataset+'_'+filetag, file_name.split('/')[-1])
-        errfile = os.path.join(errdir, dataset+'_'+filetag, file_name.split('/')[-1])
+            logfile = os.path.join(logdir, dataset+'_'+filetag, file_name.split('/')[-1])
+            outfile = os.path.join(outdir, dataset+'_'+filetag, file_name.split('/')[-1])
+            errfile = os.path.join(errdir, dataset+'_'+filetag, file_name.split('/')[-1])
 
-        script_name = srcdir+'_'.join([dataset, filetag])+'.submit'
-        write_sh(script_name, overlist_name, file_name+'.root', logfile, outfile, errfile, dataset, filetag, SPLIT)
+            script_name = srcdir+'_'.join([dataset, filetag])+'.submit'
+            write_sh(script_name, overlist_name, file_name+'.root', logfile, outfile, errfile, dataset, filetag, SPLIT)
 
-    #print listdir
-    os.system("cp -r "+listdir+" "+config)
-    #print "creating tarball from: ", TARGET
-    os.system("tar -C "+config+"/../ -czf "+TARGET+"/config.tgz config")
+    if not COUNT:
+        #print listdir
+        os.system("cp -r "+listdir+" "+config)
+        #print "creating tarball from: ", TARGET
+        os.system("tar -C "+config+"/../ -czf "+TARGET+"/config.tgz config")
 
     submit_dir  = srcdir        
     submit_list = [os.path.join(submit_dir, f) for f in os.listdir(submit_dir) if (os.path.isfile(os.path.join(submit_dir, f)) and ('.submit' in f))]
-    n_samples   = len(submit_list)
+    #n_samples   = len(submit_list)
     total_jobs  = SPLIT * total_root_files
 
     # don't submit jobs if --dry-run is used
-    if not DRY_RUN:
+    if not DRY_RUN or not COUNT:
         for f in submit_list:
             print "submitting: {0}".format(f)
             os.system('condor_submit ' + f)
